wire: force move of Peer due to casing
