+++ packages/lntools-wire/__tests__/gossip/gossip-manager.spec.ts
@@ -11,7 +11,7 @@ import { ChannelAnnouncementMessage } from "../../lib/messages/channel-announcem
 import { ChannelUpdateMessage } from "../../lib/messages/channel-update-message";
 import { ExtendedChannelAnnouncementMessage } from "../../lib/messages/extended-channel-announcement-message";
 import { NodeAnnouncementMessage } from "../../lib/messages/node-announcement-message";
-import { Peer } from "../../lib/peer";
+import { Peer } from "../../lib/Peer";
 import { PeerState } from "../../lib/PeerState";
 import { ShortChannelId } from "../../lib/shortchanid";
 import { createFakeLogger, createFakePeer } from "../_test-utils";
